fix: rename test/admin.spec.js
